refactor(web): extract socket creation
